Reworking tags and transformations, stage 3: - everything is converted to the new shiny way of doing things.   Well, everything except Gappa, which seems very unifinished anyway,   and Encoding_instantiate, which is too complex and would like to   update it with François.
Also, I commented a little piece of exception reporting in manager/,
will see it with Claude.

THIS IS STILL A WORK IN PROGRESS!
Please inform me about any bugs, ugly APIs, and proposed corrections.

All the non-implemented things, mentioned in the previous commit
message are still in the TODO list and will be done soon.