correct the make -j :
Multi-target is required for ocamlyacc but is counterproductive for ocamlc.
This is due to ocamldep which declare .cmi as dependencies only when
a .mli file exists.

That work one day because %.cmi was wrongly spelled %cmi, after the
spelling correction that doesn't work anymore. Until now :)

The last commit is also an improvement for -j