Avoid parenthesizing simple real numbers in Coq.
I am assuming these numbers are not negative. This assumption has always
existed for integers and the lexer cannot create negative real numbers
anyway. So, as long as a transformation does not create negative numbers,
there won't be any issue. (Transformations applying a negation to positive
numbers are fine though.)