fix smtv2 printer
also, remove all customization from the Z3 3.x section in prover detection.
I find it morally wrong to chase magic options of a closed-source program.
We should add an option for a prover only when we have clear understanding
of what it does and good reasons to believe that it is useful in our case.