Fix deletion of temp files on Windows 7
For KB18-007

Deleting files shortly after a subprocess used them may fail on Windows 7. We
therefore only *try* to remove it initially; if that fails, we add the
filename to a list that is cleaned up at the end of the program.  If there
still is an error deleting the file, we ignore it.