Fix incorrect detection of lablgtksourceview2 when ocamlfind is in use.
The old code was assuming that, if lablgtk2 and ocamlfind were installed,
lablgtksourceview2 was installed too. The new code explicitly checks for
both libraries.

Presumably, it won't work correctly if someone has decided to install
lablgtksourceview2 in a different location from lablgtk2, but it wouldn't
have worked with the previous version either, so it isn't a regression.