new session
Split session in two :
Session : an API for managing session without running provers
Session_scheduler : an API for running provers asynchronously

All the global states have been removed.

A session must be first read, which give a session without task.
Afterward it must be updated to the current state of the files with
some environnement and configuration.

printer and iterator are provided for session.

Session_tools : some useful functions on session.

Smoke detector : not anymore integrated to session. Just add the
      transformation "smoke_detector_top" or "smoke_detector_deep" to
      all the valid proof attempt.

prover_id are not yet removed but all is in place in session for that.