Create separate "editor" sections to support several Coq editors.
For now, there is no autodetection of the editor. The purpose of this patch
is only to provide the user with sane defaults and to avoid reusing prover
options as editor options. So the user has to manually edit the why3.conf
file to modify a prover/editor association.

Tested with Coqide 8.3 and Proof General 4.1.