Improve the heuristic for performing substitutions in the Gappa driver.
The original idea was to replace t1 by t2 only when t1 is more complicated
than t2, which was implemented by saying that t1 is neither a variable nor
a constant. Unfortunately, it means that no substitution happens when both
t1 and t2 are variables and/or constants. This patch fixes it.