whyml: do not count the used symbols in program declarations
It is difficult to count all the used symbols in programs.
On the other hand, the pure API already cares about the missing
symbols in verification conditions. So we relieve the program API
of this duty, and trust typechecking (and any other API user) to
not forget to add all the necessary dependencies to a module.

The worst possible outcome, it seems, is that one can extract
an ill-formed OCaml code if the WhyML module is done via the API,
and VCs were not generated.