Remove obsolete code for previous pretty printing mechanism
This code was used previously (GNATprove GPL) to print the precise messages
for unproved VCs. It is also the code that broke for the customer in this TN.
The underlying bug (probably in relocatable.ml) has long been fixed, but this
code is unused now so should be removed.

For L801-013