Fix some build issues when theories/plugins are obsolete or not yet installed.
The empty string is now a recognized configuration filename. It means that
the default configuration with an empty loadpath should be loaded. It also
means it will not be saved.

This removes the need for a WHY3NOCONFIG environment variable and some
tricks based on the WHY3LOADPATH variable. This also removes the need for
a /dev/null file.