Prevent time limits to grow out of control and do not update them behind the user's back.
With the previous code, one could start with an original time limit of a
few seconds and end up with proof attempts running for several minutes
after a few unsuccessful tries with provers that are aware of the actual
limit. Now, the actual time limit is never more than twice the user-defined
one.