M130-033 explode record fields in Why3 transformation
Why3 applies a transformation "eval_match" (which is essential) that
explodes records when they are quantified:

  type t = { a : int; b : int }

    forall x : t. f

  becomes
    forall a : int; b : int. let x = {a = a; b = b} in f

The problem is that this transformation does not do this for global constants

  function x : t

and this difference between quantified variables and global variables has all
kinds of strange effects (e.g. on triggers).

The new transformation in this commit does the same "explosion" for such
global constants:

  function a : int
  function b : int

(replace x with {a = a; b = b} everywhere)

rec_logic.ml
   new transformation to achieve the above
alt_ergo_bare.drv:
  activate the new transformation