upgrade Alt-Ergo proofs to 0.95.1
in three cases, we use a proof by an earlier version of Alt-Ergo
as a sample for bisection and then reprove it with other provers.

Also, moved all sessions in theories/ and modules/ to
tests/theory-sessions/. These sessions are not replayed
by nightly-builds and should not be distributed.
Please, do not run why3ide on the standard library files,
make separate files in examples/ using those theories, or
just realize them in Coq. Don't clobber the loadpath.