mlw_typing: add tuple symbols to module_uc early, after stage 1
This fixes invariant computation for a program symbol whose signature
contains a new tuple symbol. The early call of flush_tuples will miss
the tuple symbols inside specifications, but these do not participate
in invariant computation, and are not a worry. We still have to use
add_pdecl_with_tuples to catch them, though.