Termcode: add (commented out) optimistic pairing
Certain modifications in Why3 change tasks but not the propositional
structure of the goal. In this case, it is sometimes easier to use
"optimistic pairing": as long as the number of subgoals is the same
as in the old session, match them in the existing order, without
comparing shapes. This should be used only for repairing sessions
after changes in Why3.