WhyML: introduce "val ... in ..." construction
This is a syntactic sugar for higher-order "any", for which
I can't find reasonably unambiguous syntax. One can write
  val x : int <spec> in <expr>
or
  val f (x : int) : int <spec> in <expr>
or
  val f <spec> (x : int) : int <spec> in <expr>
for a function which is created via some effectful computation.
This is a generalized form of top-level "val" which only admits
latent effects (as in the second form above).