[L404-021] Save Why3 session on Ctrl-C by user
When the user stops gnatprove with Ctrl-C, we'd rather have the Why3 session
file saved on disk, so that all verification results so far are preserved.
This works out of the box for interruptions on the command line.
A special trick is needed to make it work inside GPS, as some other signal
is sent after the handler is called, so the handler starts to disable
possible SIGINT, SIGHUP and SIGTERM.

* src/gnat/gnat_main.ml
Set handler for SIGINT that save the Why3 session.