Mlw_dexpr: ignore 'old when checking user-supplied read effects
Strictly speaking, 'old should not even be a variable, we only do
it to make Mlw_wp.old_mark efficient (compared to replacement of
a constant). It is possible that presence of 'old in specs will
bite us again in the future. TODO: think about a robust way to
not count 'old among the free pvsymbols of a spec.