N414-026 Handling manual provers files
When gnatwhy3 is called with manual prover, if no proof attempt on that
goal with such prover exists then gnatwhy3 creates a new file that should
be edited by the user; else if the goal has never been proved it replays
the proof using the file provided created the first time gnatwhy3 was
called with the same manual prover.

* src/driver/call_provers.ml
(prove_file_server): added optional 'inplace' parameter to indicate whether
the file we're using is temporary or permanent (i.e. for manual proof).
Setting new save_data record field 'is_temporary'.
(handle_answer): remove vc_file only when it is temporary.

* src/driver/driver.ml
(prove_task_server): added optional 'inplace' and 'old' parameter having
the same meaning as prove_task (old = old file, inplace = do treaments on
original file). Added a distinct case for manual prover in which no new
file is created nor written.

* src/gnat/gnat_main.ml
(schedule_goal): since we don't handle when the file has been edited or
not, in the case of manual provers we always actually schedule unless it
has already been proved.
(handle_obj): when using manual provers, if it's the first time we use
the current prover on the current goal then create it's vc_file without
scheduling it.

* src/gnat/gnat_objectives.ml
(is_new_manual_proof): new function to tell if the current prover is manual
and we have never used it on given goal.
(find_goal_theory): new function to retrieve the parent theory of a goal
(create_prover_file goal): new function to create a file for the given goal
with the current prover.

* src/gnat/gnat_sched.ml
(run_goal): added code to retrieve the filename created for current prover
if it is manual and pass it to prove_task_server.
(handle_finished_call): added code to save the name of a manual prover
file.