N414-026 Reporting manual prover files' names
When an error occurs during the proof of a VC with a manual prover, the
name of the file that should contain the proof is added to the error
message. When we create a new file to manually prove a VC, an error occurs
so we learn the name of the created file.

* src/gnat/gnat_main.ml
(handle_vc_result): provide potential filename to report when vc is not
proved.
(handle_obj): register message when we create a file

* src/gnat/gnat_objectives.ml
Moved is_new_manual_proof, find_goal_theory and create_prover_file to new
file: src/gnat/gnat_manual.ml

* src/gnat/gnat_manual.ml
New file to group functions relative to manual provers.
(get_prover_file): function to get the manual prover file of a given VC.

* src/gnat/gnat_report.ml
Added 'vc_file' optional field to msg record
(register): added 'filename' optional parameter.
(print_msg): print the 'vc_file' field when present.
(print_json_msg): add 'vc_file' field to the JSON record.