N414-026 Fix incomplete proofs
If a user modifies Why3 generated code for manual provers, then he might
get proven VC which are incomplete or not corresponding to original VC.
We have to rewrite the VC file each time the VC is scheduled.

* src/gnat/gnat_manual.ml:
(rewrite_goal): new function to rewrite a VC file corresponding to a
given goal.

*src/gnat/gnat_main.ml:
(actually_schedule_goal): rewrite file when we are using a manual prover.