N611-025 Generating proof filenames with SPARK code "shapes"
This modification changes the names generated for files used in
interactive proofs.

* src/gnat/gnat_expl.ml
Modified types "check" and "gp_label" to handle a field/variant for
the shape.

* src/gnat/gnat_manual.ml
(resize_shape): new function, used to reduce the shape if it would
cause the generation of a filename too big for the filesystem.
(compute_filename): compute the name from the why module name +
the shape.