Hashcons: create unique objects
this may be useful if we want to have in an h-consed type
objects that cannot be equal to no other existing or future
value of that type (thus, this objects cannot be reconstructed).
This is more efficient than making comparison return false on
these objects, since we don't need to store them in a week set
at all.