New transformations induction_pr and inversion_pr.
Induction_pr performs induction on the leftmost application
of some inductive predicate in the goal.

This can be overriden by attaching the label "induction"
on such an application. In this case, any unlabeled application
will be ignored.

Inversion_pr works similarly, excepted that it does not generate
inductive hypotheses, but simply inverses inductive predicate.
In this case, the label "inversion" will be used.