N902-018 allow list of provers
The option --prover=string has been modified. Now the string can be a
comma-separated list of prover names, which will all be tried in that
order on a given VC. So all of these work and do different things

  --prover=altergo
  --prover=cvc4
  --prover=altergo,cvc4
  --prover=cvc4,altergo

Manual provers currently cannot be combined, so

  --prover=coq

but not this:

  --prover=altergo,coq

(could also make sense, e.g. try first to prove automatically, then
generate coq VCs)

* gnat_config.ml
  parse the --prover option differently, load drivers for all provers,
  sanity check that only one manual prover is present
* gnat_sched.ml
  scheduler now needs to know which prover to run, and stores that info
  in the queue
* gnat_report.ml
  slight restructuring to suppress the need to know "the" global prover
* gnat_manual.ml
  change the way "the" manual prover is accessed
* gnat_main.ml
  changes mostly to handling of manual prover case, and passing info
  around e.g. to gnat_manual
* gnat_objectives.ml
  now reschedule a goal if there is a prover left to try