Initial implementation of counter-examples.
If all provers fail to prove a VC, additional prover run is scheduled to
get a counter-example. This prover run is made using cvc4.
Note that cvc4 is run with different arguments than in case of standard
run and its input can also differ (there are some why3 transformations
done just in the case of generating counter-examples).

A counter-example is written to the file in gnatprove directory. This file
is reported in standard output (the same way as, e.g., the file with
counter-example trace).