O622-007 Use steps support of why3
The number of steps was passed to provers by writing the number to the
why.conf file generated by gnatprove. Additionally, step number needed
to be passed to why3 (why3 needs to know whether steps were given or not).
This commit removes the code for handling steps from gnatprove. Steps
are now just passed to gnatwhy3 and the support of steps built in why3
is used.

Added the code for adjusting the steplimit for CVC4, added code for using
the version of the command for the prover for steplimit if steplimit is
given. Removed the code used to parse steplimit from the output of the
prover in gnatwhy3 - steplimit is already parsed by why3 before and passed
in prover_result.

* src/gnat/gnat_config.ml
(steps): Added a prover as a parameter. This parameter is used to decide
how to adjust the steps value.

* src/gnat/gnat_sched.ml
(run_goal): Getting right version of the prover command (with or without
steps), passing adjusted steplimit information.

* src/gnat/gnat_report.ml
Removed code for getting steplimit from the output of the prover, using
steplimit from prover_result instead.